OK-38

+++ app/Http/Controllers/CurriculumController.php
@@ -12,44 +12,50 @@ class CurriculumController extends Controller
 {
     public function index(Request $request)
     {
-        $query = Curriculum::with(['department', 'specialty', 'edu_year']);
-        if ($request->filled('department_id')) {
-            $query->where('department_id', $request->department_id);
-        }
-        if ($request->filled('specialty_id')) {
-            $query->where('specialty_id', $request->specialty_id);
-        }
-        if ($request->filled('edu_year_id')) {
-            $query->where('edu_year_id', $request->edu_year_id);
-        }
-        $curr = $query->orderBy('id', 'desc')->paginate(20);
-        $departments = Department::where('structure', '11')->orderBy('name')->get();
-        if ($request->filled('department_id')) {
-            $specialties = Specialty::where('department_id', $request->department_id)->get();
-        } else {
-            $specialties = Specialty::all();
-        }
-        $eduYears = EduYear::all();
-        return view('pages.web.curriculum.index', compact('curr', 'departments', 'specialties', 'eduYears'));
+        if (auth()->user()->can('curriculum.view')) {
+            $query = Curriculum::with(['department', 'specialty', 'edu_year']);
+            if ($request->filled('department_id')) {
+                $query->where('department_id', $request->department_id);
+            }
+            if ($request->filled('specialty_id')) {
+                $query->where('specialty_id', $request->specialty_id);
+            }
+            if ($request->filled('edu_year_id')) {
+                $query->where('edu_year_id', $request->edu_year_id);
+            }
+            $curr = $query->orderBy('id', 'desc')->paginate(20);
+            $departments = Department::where('structure', '11')->orderBy('name')->get();
+            if ($request->filled('department_id')) {
+                $specialties = Specialty::where('department_id', $request->department_id)->get();
+            } else {
+                $specialties = Specialty::all();
+            }
+            $eduYears = EduYear::all();
+            return view('pages.web.curriculum.index', compact('curr', 'departments', 'specialties', 'eduYears'));
+        } else abort(404);
     }
 
     public function destroy($id)
     {
-        $curriculum = Curriculum::findOrFail($id);
-        if ($curriculum->subjects()->exists()) {
-            return redirect()->back()->with('error', 'Bu o‘quv rejada fanlar mavjud! Oldin fanlarni o‘chiring yoki boshqa rejaga o‘tkazing.');
+        if (auth()->user()->can('curriculum.delete')) {
+            $curriculum = Curriculum::findOrFail($id);
+            if ($curriculum->subjects()->exists()) {
+                return redirect()->back()->with('error', 'Bu o‘quv rejada fanlar mavjud! Oldin fanlarni o‘chiring yoki boshqa rejaga o‘tkazing.');
+            }
+            $curriculum->delete();
+            return redirect()->back()->with('success', 'O‘quv reja muvaffaqiyatli o‘chirildi!');
         }
-        $curriculum->delete();
-        return redirect()->back()->with('success', 'O‘quv reja muvaffaqiyatli o‘chirildi!');
     }
 
     public function getSpecialties(Request $request)
     {
-        if ($request->has('department_id') && $request->department_id != null) {
-            $specialties = Specialty::where('department_id', $request->department_id)->get();
-        } else {
-            $specialties = Specialty::all();
+        if (auth()->user()->can('curriculum.view')) {
+            if ($request->has('department_id') && $request->department_id != null) {
+                $specialties = Specialty::where('department_id', $request->department_id)->get();
+            } else {
+                $specialties = Specialty::all();
+            }
+            return response()->json($specialties);
         }
-        return response()->json($specialties);
     }
 }
